+++ Gestion Auto-Ecole 2015/Gestion Auto-École/SharpZipLib/src/GZip/GzipOutputStream.cs
@@ -235,10 +235,10 @@ namespace ICSharpCode.SharpZipLib.GZip
 				int mod_time = (int)((DateTime.Now.Ticks - new DateTime(1970, 1, 1).Ticks) / 10000000L);  // Ticks give back 100ns intervals
 				byte[] gzipHeader = {
 					// The two magic bytes
-					(byte) (GZipConstants.GZIP_MAGIC >> 8), (byte) (GZipConstants.GZIP_MAGIC & 0xff),
+					 GZipConstants.GZIP_MAGIC >> 8,  GZipConstants.GZIP_MAGIC & 0xff,
 
 					// The compression type
-					(byte) Deflater.DEFLATED,
+					 Deflater.DEFLATED,
 
 					// The flags (not set)
 					0,
@@ -251,8 +251,8 @@ namespace ICSharpCode.SharpZipLib.GZip
 					0,
 
 					// The OS type (unknown)
-					(byte) 255
-				};
+					 255
+                };
 				baseOutputStream_.Write(gzipHeader, 0, gzipHeader.Length);
 			}
 		}
