+++ Gestion Auto-Ecole 2015/Gestion Auto-École/SharpZipLib/src/Zip/Compression/Streams/InflaterInputStream.cs
@@ -210,7 +210,7 @@ namespace ICSharpCode.SharpZipLib.Zip.Compression.Streams
 					}
 				}
 				int toCopy = Math.Min(currentLength, available);
-				System.Array.Copy(rawData, rawLength - (int)available, outBuffer, currentOffset, toCopy);
+				System.Array.Copy(rawData, rawLength - available, outBuffer, currentOffset, toCopy);
 				currentOffset += toCopy;
 				currentLength -= toCopy;
 				available -= toCopy;
@@ -243,7 +243,7 @@ namespace ICSharpCode.SharpZipLib.Zip.Compression.Streams
 				}
 				
 				int toCopy = Math.Min(currentLength, available);
-				Array.Copy(clearText, clearTextLength - (int)available, outBuffer, currentOffset, toCopy);
+				Array.Copy(clearText, clearTextLength - available, outBuffer, currentOffset, toCopy);
 				currentOffset += toCopy;
 				currentLength -= toCopy;
 				available -= toCopy;
